added more server side tests
