+++ package.json
@@ -5,7 +5,7 @@
   "main": "/server/server.js",
   "scripts": {
     "start": "./node_modules/.bin/babel-node ./server/app.js",
-    "test": "NODE_ENV=test mocha --compilers js:babel-core/register server/test/test.js --timeout 10000",
+    "test": "NODE_ENV=test mocha --compilers js:babel-core/register server/test --timeout 10000",
     "build": "babel server -d dist",
     "start:dev": "nodemon --watch server --exec babel-node -- ./server/bin/www",
     "dev": "nodemon server/server.js & webpack-dev-server",
